Add event data and retrieval function to action.js
