+++ app/events/action.js
@@ -1,0 +1,176 @@
+"use server";
+
+const events = [
+  {
+    eventPoster: "EventDemo1.svg",
+    eventName: "KODESPHERE - The Flagship Event",
+    eventLocation: "Campus-13",
+    eventDate: "Sat-29-Mar-2025",
+    eventPrize: 200.0,
+  },
+  {
+    eventPoster: "EventDemo2.svg",
+    eventName: "SPOTLIGHT - The Flagship Event",
+    eventLocation: "Campus-14",
+    eventDate: "Sat-25-Mar-2025",
+    eventPrize: 300.0,
+  },
+  {
+    eventPoster: "EventDemo3.svg",
+    eventName: "STEP IT UP - The Flagship Event",
+    eventLocation: "Campus-15",
+    eventDate: "Sat-25-Mar-2025",
+    eventPrize: 400.0,
+  },
+  {
+    eventPoster: "EventDemo1.svg",
+    eventName: "KODESPHERE - The Flagship Event",
+    eventLocation: "Campus-13",
+    eventDate: "Sat-29-Mar-2025",
+    eventPrize: 200.0,
+  },
+  {
+    eventPoster: "EventDemo2.svg",
+    eventName: "SPOTLIGHT - The Flagship Event",
+    eventLocation: "Campus-14",
+    eventDate: "Sat-25-Mar-2025",
+    eventPrize: 300.0,
+  },
+  {
+    eventPoster: "EventDemo3.svg",
+    eventName: "STEP IT UP - The Flagship Event",
+    eventLocation: "Campus-15",
+    eventDate: "Sat-25-Mar-2025",
+    eventPrize: 400.0,
+  },
+  {
+    eventPoster: "EventDemo1.svg",
+    eventName: "KODESPHERE - The Flagship Event",
+    eventLocation: "Campus-13",
+    eventDate: "Sat-29-Mar-2025",
+    eventPrize: 200.0,
+  },
+  {
+    eventPoster: "EventDemo2.svg",
+    eventName: "SPOTLIGHT - The Flagship Event",
+    eventLocation: "Campus-14",
+    eventDate: "Sat-25-Mar-2025",
+    eventPrize: 300.0,
+  },
+  {
+    eventPoster: "EventDemo3.svg",
+    eventName: "STEP IT UP - The Flagship Event",
+    eventLocation: "Campus-15",
+    eventDate: "Sat-25-Mar-2025",
+    eventPrize: 400.0,
+  },
+  {
+    eventPoster: "EventDemo1.svg",
+    eventName: "KODESPHERE - The Flagship Event",
+    eventLocation: "Campus-13",
+    eventDate: "Sat-29-Mar-2025",
+    eventPrize: 200.0,
+  },
+  {
+    eventPoster: "EventDemo2.svg",
+    eventName: "SPOTLIGHT - The Flagship Event",
+    eventLocation: "Campus-14",
+    eventDate: "Sat-25-Mar-2025",
+    eventPrize: 300.0,
+  },
+  {
+    eventPoster: "EventDemo3.svg",
+    eventName: "STEP IT UP - The Flagship Event",
+    eventLocation: "Campus-15",
+    eventDate: "Sat-25-Mar-2025",
+    eventPrize: 400.0,
+  },
+  {
+    eventPoster: "EventDemo1.svg",
+    eventName: "KODESPHERE - The Flagship Event",
+    eventLocation: "Campus-13",
+    eventDate: "Sat-29-Mar-2025",
+    eventPrize: 200.0,
+  },
+  {
+    eventPoster: "EventDemo2.svg",
+    eventName: "SPOTLIGHT - The Flagship Event",
+    eventLocation: "Campus-14",
+    eventDate: "Sat-25-Mar-2025",
+    eventPrize: 300.0,
+  },
+  {
+    eventPoster: "EventDemo3.svg",
+    eventName: "STEP IT UP - The Flagship Event",
+    eventLocation: "Campus-15",
+    eventDate: "Sat-25-Mar-2025",
+    eventPrize: 400.0,
+  },
+  {
+    eventPoster: "EventDemo1.svg",
+    eventName: "KODESPHERE - The Flagship Event",
+    eventLocation: "Campus-13",
+    eventDate: "Sat-29-Mar-2025",
+    eventPrize: 200.0,
+  },
+  {
+    eventPoster: "EventDemo2.svg",
+    eventName: "SPOTLIGHT - The Flagship Event",
+    eventLocation: "Campus-14",
+    eventDate: "Sat-25-Mar-2025",
+    eventPrize: 300.0,
+  },
+  {
+    eventPoster: "EventDemo3.svg",
+    eventName: "STEP IT UP - The Flagship Event",
+    eventLocation: "Campus-15",
+    eventDate: "Sat-25-Mar-2025",
+    eventPrize: 400.0,
+  },
+  {
+    eventPoster: "EventDemo1.svg",
+    eventName: "KODESPHERE - The Flagship Event",
+    eventLocation: "Campus-13",
+    eventDate: "Sat-29-Mar-2025",
+    eventPrize: 200.0,
+  },
+  {
+    eventPoster: "EventDemo2.svg",
+    eventName: "SPOTLIGHT - The Flagship Event",
+    eventLocation: "Campus-14",
+    eventDate: "Sat-25-Mar-2025",
+    eventPrize: 300.0,
+  },
+  {
+    eventPoster: "EventDemo3.svg",
+    eventName: "STEP IT UP - The Flagship Event",
+    eventLocation: "Campus-15",
+    eventDate: "Sat-25-Mar-2025",
+    eventPrize: 400.0,
+  },
+  {
+    eventPoster: "EventDemo1.svg",
+    eventName: "KODESPHERE - The Flagship Event",
+    eventLocation: "Campus-13",
+    eventDate: "Sat-29-Mar-2025",
+    eventPrize: 200.0,
+  },
+  {
+    eventPoster: "EventDemo2.svg",
+    eventName: "SPOTLIGHT - The Flagship Event",
+    eventLocation: "Campus-14",
+    eventDate: "Sat-25-Mar-2025",
+    eventPrize: 300.0,
+  },
+  {
+    eventPoster: "EventDemo3.svg",
+    eventName: "STEP IT UP - The Flagship Event",
+    eventLocation: "Campus-15",
+    eventDate: "Sat-25-Mar-2025",
+    eventPrize: 400.0,
+  },
+];
+
+export async function getEvents() {
+  return { data: events };
+}
